auto

+++ src/api/controllers/gen/offer-amounts.controller.ts
@@ -11,58 +11,75 @@ import { ApiKeys } from './../../../entities/ApiKeys'
 import { CreateOfferAmountDto } from './../../dtos/create-offer-amount.dto'
 import { UpdateOfferAmountDto } from './../../dtos/update-offer-amount.dto'
 import { CatalogProductOffersService } from './../../services/catalog-product-offers.service'
+import { CatalogsService } from './../../services/catalogs.service'
 import { OfferAmountsService } from './../../services/offer-amounts.service'
 import { StoresService } from './../../services/stores.service'
 import { EntityManager } from '@mikro-orm/postgresql'
-import { Controller, HttpException, HttpStatus, UseGuards } from '@nestjs/common'
+import { ClassSerializerInterceptor, Controller, HttpException, HttpStatus, UseGuards, UseInterceptors } from '@nestjs/common'
 import { AuthGuard } from '@nestjs/passport'
 import { ApiHeader, ApiTags } from '@nestjs/swagger'
 
 @ApiHeader({ name: 'X-API-KEY', required: true })
 @UseGuards(AuthGuard('api-key'))
 @ApiTags('Offer amounts')
-@Controller('catalog/:catalog/offer-amount')
+@Controller('catalog/:catalog/offer/:offer/store')
+@UseInterceptors(ClassSerializerInterceptor)
 export class GenOfferAmountsController {
 	constructor(
 		protected readonly catalogProductOffersService: CatalogProductOffersService,
+		protected readonly catalogsService: CatalogsService,
 		protected readonly offerAmountsService: OfferAmountsService,
 		protected readonly storesService: StoresService,
 	) { }
 	
-	async findAll(apiKey: ApiKeys, offset: number, limit: number) {
+	async findAll(apiKey: ApiKeys, catalog: number, offset: number, limit: number) {
+		const catalogIns0 = await this.catalogsService.findById(catalog);
+		if(catalogIns0===null || !(catalogIns0.company.id===apiKey.company.id)){
+			throw new HttpException('Catalog not found', HttpStatus.NOT_FOUND);
+		}
 		if(offset<0) throw new HttpException('Wrong offset value', HttpStatus.BAD_REQUEST);
 		if(limit<0) throw new HttpException('Wrong limit value', HttpStatus.BAD_REQUEST);
 		if(limit>1000) limit = 1000; // throw new HttpException('Wrong limit value', HttpStatus.BAD_REQUEST);
 		return await this.offerAmountsService.listAll(offset, limit);
 	}
 	
-	async findOne(apiKey: ApiKeys, offer: bigint, store: number) {
+	async findOne(apiKey: ApiKeys, catalog: number, offer: bigint, store: number) {
+		const catalogIns0 = await this.catalogsService.findById(catalog);
+		if(catalogIns0===null || !(catalogIns0.company.id===apiKey.company.id)){
+			throw new HttpException('Catalog not found', HttpStatus.NOT_FOUND);
+		}
 		const entity = await this.offerAmountsService.findByOfferAndStore(offer, store);
 		if(entity===null){
 			throw new HttpException('Entity not found', HttpStatus.NOT_FOUND);
 		}
-		await this.validateRead(entity, apiKey, offer, store);
+		await this.validateRead(entity, apiKey, catalog, offer, store);
 		return entity;
 	}
 	
-	async validateRead(entity, apiKey: ApiKeys, offer: bigint, store: number) { }
+	async validateRead(entity, apiKey: ApiKeys, catalog: number, offer: bigint, store: number) { }
 	
-	async update(apiKey: ApiKeys, offer: bigint, store: number, updateDto: UpdateOfferAmountDto) {
+	async update(apiKey: ApiKeys, catalog: number, offer: bigint, store: number, updateDto: UpdateOfferAmountDto) {
+		updateDto.offer = offer;
+		updateDto.store = store;
+		const catalogIns0 = await this.catalogsService.findById(catalog);
+		if(catalogIns0===null || !(catalogIns0.company.id===apiKey.company.id)){
+			throw new HttpException('Catalog not found', HttpStatus.NOT_FOUND);
+		}
 		return await this.offerAmountsService.transactional(async (em) => {
 			const entity = await this.offerAmountsService.findByOfferAndStore(offer, store, em);
 			if((updateDto.store!==undefined && updateDto.store!==entity.store.id)){
-				const tmp0 = await this.storesService.findById(updateDto.store, em);
-				if(tmp0===null){
-					throw new HttpException('Not found contrainst (store)', HttpStatus.CONFLICT);
+				const storeIns1 = await this.storesService.findById(updateDto.store);
+				if(storeIns1===null || !(storeIns1.company.id===apiKey.company.id)){
+					throw new HttpException('Store not found', HttpStatus.NOT_FOUND);
 				}
 			}
 			if((updateDto.offer!==undefined && updateDto.offer!==entity.offer.id)){
-				const tmp1 = await this.catalogProductOffersService.findById(updateDto.offer, em);
-				if(tmp1===null){
-					throw new HttpException('Not found contrainst (offer)', HttpStatus.CONFLICT);
+				const offerIns2 = await this.catalogProductOffersService.findById(updateDto.offer);
+				if(offerIns2===null || !(offerIns2.product.catalog.id===catalog)){
+					throw new HttpException('Offer not found', HttpStatus.NOT_FOUND);
 				}
 			}
-			await this.validateUpdate(entity, apiKey, offer, store, updateDto, em);
+			await this.validateUpdate(entity, apiKey, catalog, offer, store, updateDto, em);
 			if(entity===null){
 				return await this.offerAmountsService.update(entity, updateDto, em);
 			} else {
@@ -71,19 +88,23 @@ export class GenOfferAmountsController {
 		});
 	}
 	
-	async validateUpdate(entity, apiKey: ApiKeys, offer: bigint, store: number, updateDto: UpdateOfferAmountDto, em: EntityManager) { }
+	async validateUpdate(entity, apiKey: ApiKeys, catalog: number, offer: bigint, store: number, updateDto: UpdateOfferAmountDto, em: EntityManager) { }
 	
-	async delete(apiKey: ApiKeys, offer: bigint, store: number) {
+	async delete(apiKey: ApiKeys, catalog: number, offer: bigint, store: number) {
+		const catalogIns0 = await this.catalogsService.findById(catalog);
+		if(catalogIns0===null || !(catalogIns0.company.id===apiKey.company.id)){
+			throw new HttpException('Catalog not found', HttpStatus.NOT_FOUND);
+		}
 		return await this.offerAmountsService.transactional(async (em) => {
 			const entity = await this.offerAmountsService.findByOfferAndStore(offer, store, em);
 			if(entity===null){
 				throw new HttpException('Entity not found', HttpStatus.NOT_FOUND);
 			}
-			await this.validateDelete(entity, apiKey, offer, store, em);
+			await this.validateDelete(entity, apiKey, catalog, offer, store, em);
 			return await this.offerAmountsService.remove(entity, em);
 		});
 	}
 	
-	async validateDelete(entity, apiKey: ApiKeys, offer: bigint, store: number, em: EntityManager) { }
+	async validateDelete(entity, apiKey: ApiKeys, catalog: number, offer: bigint, store: number, em: EntityManager) { }
 	
 }
